requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
Pillow==10.1.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.28.64
pytz==2023.3.post1
python-decouple==3.8